+++ src/main/java/nextstep/subway/path/domain/PathFinder.java
@@ -1,7 +1,9 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.path.domain;
 
 import nextstep.subway.exception.NotConnectStationException;
-import nextstep.subway.line.dto.PathResponse;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.domain.Section;
+import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.domain.Station;
 import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
