+++ src/main/java/nextstep/subway/path/ui/PathController.java
@@ -1,7 +1,7 @@
-package nextstep.subway.line.ui;
+package nextstep.subway.path.ui;
 
-import nextstep.subway.line.application.PathService;
-import nextstep.subway.line.dto.PathResponse;
+import nextstep.subway.path.application.PathService;
+import nextstep.subway.path.dto.PathResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
